requirements.txt
openai>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
sse-starlette>=2.0.0
python-multipart>=0.0.9
